requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Security & Encryption
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & Async
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0
sqlalchemy==2.0.23

# Redis
redis==5.0.1
hiredis==2.3.2

# External Services (will be needed soon)
twilio==8.11.0
# deepgram-sdk==3.0.0  # Uncomment when implementing STT
# anthropic==0.7.8  # Uncomment when implementing Claude
# elevenlabs==0.2.27  # Uncomment when implementing TTS
# google-auth==2.25.2  # Uncomment when implementing Calendar
# google-auth-oauthlib==1.2.0
# google-auth-httplib2==0.1.1
# google-api-python-client==2.111.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilities
tenacity==8.2.3
pytz==2023.3.post1
email-validator==2.1.0  # Optional: For advanced international email validation
numpy==1.26.2  # For audio processing and VAD

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.4.0
faker==20.1.0
locust==2.17.0  # For load testing

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Production Server (uncomment for production)
# gunicorn==21.2.0

# Note: For production deployment, consider using pip-tools
# to generate pinned requirements with pip-compile